+++ HDL/instr_dec.v
@@ -150,6 +150,10 @@ module instr_dec
 	parameter  F_SU2   = 0,   // floating-point subtraction at input 2
     parameter SF_SU1   = 0,   // floating-point subtraction at input 1 with stack
     parameter SF_SU2   = 0,   // floating-point subtraction at input 2 with stack
+    parameter  F_SCL   = 0,   // scale float by 2^k, k from memory
+    parameter SF_SCL   = 0,   // scale float by 2^k, k from stack
+    parameter    XPO   = 0,   // base-2 exponent of float (acc) as int
+    parameter  XPO_M   = 0,   // base-2 exponent of float (memory) as int
 
     // base-less indirect addressing (runtime pointers / array params)
     parameter    LDA   = 0,   // acc = mem[acc]
@@ -325,6 +329,13 @@ wire  wF_SU2  ; generate if ((F_SU2) != 0) begin : dec_wf_su2 assign wF_SU2 = op
 wire wSF_SU1  ; generate if ((SF_SU1) != 0) begin : dec_wsf_su1 assign wSF_SU1 = opcode == 7'd100; end else begin : dec_wsf_su1 assign wSF_SU1 = 1'b0; end endgenerate
 wire wSF_SU2  ; generate if ((SF_SU2) != 0) begin : dec_wsf_su2 assign wSF_SU2 = opcode == 7'd101; end else begin : dec_wsf_su2 assign wSF_SU2 = 1'b0; end endgenerate
 
+// scale-by-2^k and exponent-extract ------------------------------------------
+
+wire  wF_SCL  ; generate if ((F_SCL ) != 0) begin : dec_wf_scl  assign wF_SCL  = opcode == 7'd104; end else begin : dec_wf_scl  assign wF_SCL  = 1'b0; end endgenerate
+wire wSF_SCL  ; generate if ((SF_SCL) != 0) begin : dec_wsf_scl assign wSF_SCL = opcode == 7'd105; end else begin : dec_wsf_scl assign wSF_SCL = 1'b0; end endgenerate
+wire    wXPO  ; generate if ((XPO   ) != 0) begin : dec_wxpo    assign wXPO    = opcode == 7'd106; end else begin : dec_wxpo    assign wXPO    = 1'b0; end endgenerate
+wire  wXPO_M  ; generate if ((XPO_M ) != 0) begin : dec_wxpo_m  assign wXPO_M  = opcode == 7'd107; end else begin : dec_wxpo_m  assign wXPO_M  = 1'b0; end endgenerate
+
 // base-less indirect addressing ----------------------------------------------
 
 wire    wLDA  ; generate if ((LDA) != 0) begin : dec_wlda assign wLDA = opcode == 7'd102; end else begin : dec_wlda assign wLDA = 1'b0; end endgenerate
@@ -363,17 +374,17 @@ generate if (P_LOD | PSH | P_INN | PF_INN | P_NEG_M | PF_NEG_M | P_ABS_M | PF_AB
 
 // data-stack read control circuit --------------------------------------------
 
-generate if (SET_P | STI | ISI | POP | S_ADD | SF_ADD | S_MLT | SF_MLT | S_DIV | SF_DIV | S_MOD | S_SGN | SF_SGN | S_AND | S_ORR | S_XOR | S_LAN | S_LOR | S_LES | SF_LES | S_GRE | SF_GRE | S_EQU | S_SHL | S_SHR | S_SRS | SF_SU1 | SF_SU2 | STA) begin : dec_pop assign pop = wSET_P | wSTI | wISI | wPOP | wS_ADD | wSF_ADD | wS_MLT | wSF_MLT | wS_DIV | wSF_DIV | wS_MOD | wS_SGN | wSF_SGN | wS_AND | wS_ORR | wS_XOR | wS_LAN | wS_LOR | wS_LES | wSF_LES | wS_GRE | wSF_GRE | wS_EQU | wS_SHL | wS_SHR | wS_SRS | wSF_SU1 | wSF_SU2 | wSTA; end else begin : dec_pop assign pop = 1'b0; end endgenerate
+generate if (SET_P | STI | ISI | POP | S_ADD | SF_ADD | S_MLT | SF_MLT | S_DIV | SF_DIV | S_MOD | S_SGN | SF_SGN | S_AND | S_ORR | S_XOR | S_LAN | S_LOR | S_LES | SF_LES | S_GRE | SF_GRE | S_EQU | S_SHL | S_SHR | S_SRS | SF_SU1 | SF_SU2 | SF_SCL | STA) begin : dec_pop assign pop = wSET_P | wSTI | wISI | wPOP | wS_ADD | wSF_ADD | wS_MLT | wSF_MLT | wS_DIV | wSF_DIV | wS_MOD | wS_SGN | wSF_SGN | wS_AND | wS_ORR | wS_XOR | wS_LAN | wS_LOR | wS_LES | wSF_LES | wS_GRE | wSF_GRE | wS_EQU | wS_SHL | wS_SHR | wS_SRS | wSF_SU1 | wSF_SU2 | wSF_SCL | wSTA; end else begin : dec_pop assign pop = 1'b0; end endgenerate
 
 // ALU operations control circuit ---------------------------------------------
 
 wire b5,b4,b3,b2,b1,b0;
 
 // logic for b5
-generate if (INV | INV_M | P_INV_M | LAN | S_LAN | LOR | S_LOR | LIN | LIN_M | P_LIN_M | LES | S_LES | F_LES | SF_LES | GRE | S_GRE | F_GRE | SF_GRE | EQU | S_EQU | SHL | S_SHL | SHR | S_SHR | SRS | S_SRS | F_ROT | F_SU1 | F_SU2 | SF_SU1 | SF_SU2) begin : dec_b5 assign b5 = wINV | wINV_M | wP_INV_M | wLAN | wS_LAN | wLOR | wS_LOR | wLIN | wLIN_M | wP_LIN_M | wLES | wS_LES | wF_LES | wSF_LES | wGRE | wS_GRE | wF_GRE | wSF_GRE | wEQU | wS_EQU | wSHL | wS_SHL | wSHR | wS_SHR | wSRS | wS_SRS | wF_ROT | wF_SU1 | wF_SU2 | wSF_SU1 | wSF_SU2; end else begin : dec_b5 assign b5 = 1'b0; end endgenerate
+generate if (INV | INV_M | P_INV_M | LAN | S_LAN | LOR | S_LOR | LIN | LIN_M | P_LIN_M | LES | S_LES | F_LES | SF_LES | GRE | S_GRE | F_GRE | SF_GRE | EQU | S_EQU | SHL | S_SHL | SHR | S_SHR | SRS | S_SRS | F_ROT | F_SU1 | F_SU2 | SF_SU1 | SF_SU2 | F_SCL | SF_SCL | XPO | XPO_M) begin : dec_b5 assign b5 = wINV | wINV_M | wP_INV_M | wLAN | wS_LAN | wLOR | wS_LOR | wLIN | wLIN_M | wP_LIN_M | wLES | wS_LES | wF_LES | wSF_LES | wGRE | wS_GRE | wF_GRE | wSF_GRE | wEQU | wS_EQU | wSHL | wS_SHL | wSHR | wS_SHR | wSRS | wS_SRS | wF_ROT | wF_SU1 | wF_SU2 | wSF_SU1 | wSF_SU2 | wF_SCL | wSF_SCL | wXPO | wXPO_M; end else begin : dec_b5 assign b5 = 1'b0; end endgenerate
 
 // logic for b4
-generate if (F_INN | PF_INN | ABS_M | P_ABS_M | F_ABS | F_ABS_M | PF_ABS_M | PST | PST_M | P_PST_M | F_PST | F_PST_M | PF_PST_M | NRM | NRM_M | P_NRM_M | I2F | I2F_M | P_I2F_M | F2I | F2I_M | P_F2I_M | AND | S_AND | ORR | S_ORR | XOR | S_XOR | F_SU2 | SF_SU2) begin : dec_b4 assign b4 = wF_INN | wPF_INN | wABS_M | wP_ABS_M | wF_ABS | wF_ABS_M | wPF_ABS_M | wPST | wPST_M | wP_PST_M | wF_PST | wF_PST_M | wPF_PST_M | wNRM | wNRM_M | wP_NRM_M | wI2F | wI2F_M | wP_I2F_M | wF2I | wF2I_M | wP_F2I_M | wAND | wS_AND | wORR | wS_ORR | wXOR | wS_XOR | wF_SU2 | wSF_SU2; end else begin : dec_b4 assign b4 = 1'b0; end endgenerate
+generate if (F_INN | PF_INN | ABS_M | P_ABS_M | F_ABS | F_ABS_M | PF_ABS_M | PST | PST_M | P_PST_M | F_PST | F_PST_M | PF_PST_M | NRM | NRM_M | P_NRM_M | I2F | I2F_M | P_I2F_M | F2I | F2I_M | P_F2I_M | AND | S_AND | ORR | S_ORR | XOR | S_XOR | F_SU2 | SF_SU2 | F_SCL | SF_SCL | XPO | XPO_M) begin : dec_b4 assign b4 = wF_INN | wPF_INN | wABS_M | wP_ABS_M | wF_ABS | wF_ABS_M | wPF_ABS_M | wPST | wPST_M | wP_PST_M | wF_PST | wF_PST_M | wPF_PST_M | wNRM | wNRM_M | wP_NRM_M | wI2F | wI2F_M | wP_I2F_M | wF2I | wF2I_M | wP_F2I_M | wAND | wS_AND | wORR | wS_ORR | wXOR | wS_XOR | wF_SU2 | wSF_SU2 | wF_SCL | wSF_SCL | wXPO | wXPO_M; end else begin : dec_b4 assign b4 = 1'b0; end endgenerate
 
 // logic for b3
 generate if (F_INN | PF_INN | MOD | S_MOD | SGN | S_SGN | F_SGN | SF_SGN | NEG | NEG_M | P_NEG_M | F_NEG | F_NEG_M | PF_NEG_M | ABS | NRM_M | P_NRM_M | I2F | I2F_M | P_I2F_M | F2I | F2I_M | P_F2I_M | AND | S_AND | ORR | S_ORR | XOR | S_XOR | GRE | S_GRE | F_GRE | SF_GRE | EQU | S_EQU | SHL | S_SHL | SHR | S_SHR | SRS | S_SRS | F_ROT | F_SU1 | SF_SU1) begin : dec_b3 assign b3 = wF_INN | wPF_INN | wMOD | wS_MOD | wSGN | wS_SGN | wF_SGN | wSF_SGN | wNEG | wNEG_M | wP_NEG_M | wF_NEG | wF_NEG_M | wPF_NEG_M | wABS | wNRM_M | wP_NRM_M | wI2F | wI2F_M | wP_I2F_M | wF2I | wF2I_M | wP_F2I_M | wAND | wS_AND | wORR | wS_ORR | wXOR | wS_XOR | wGRE | wS_GRE | wF_GRE | wSF_GRE | wEQU | wS_EQU | wSHL | wS_SHL | wSHR | wS_SHR | wSRS | wS_SRS | wF_ROT | wF_SU1 | wSF_SU1; end else begin : dec_b3 assign b3 = 1'b0; end endgenerate
@@ -382,10 +393,10 @@ generate if (F_INN | PF_INN | MOD | S_MOD | SGN | S_SGN | F_SGN | SF_SGN | NEG |
 generate if (MLT | S_MLT | F_MLT | SF_MLT | DIV | S_DIV | F_DIV | SF_DIV | NEG_M | P_NEG_M | F_NEG | F_NEG_M | PF_NEG_M | ABS | PST_M | P_PST_M | F_PST | F_PST_M | PF_PST_M | NRM | F2I_M | P_F2I_M | AND | S_AND | ORR | S_ORR | XOR | S_XOR | LIN | LIN_M | P_LIN_M | LES | S_LES | F_LES | SF_LES | SHR | S_SHR | SRS | S_SRS | F_ROT | F_SU1 | SF_SU1) begin : dec_b2 assign b2 = wMLT | wS_MLT | wF_MLT | wSF_MLT | wDIV | wS_DIV | wF_DIV | wSF_DIV | wNEG_M | wP_NEG_M | wF_NEG | wF_NEG_M | wPF_NEG_M | wABS | wPST_M | wP_PST_M | wF_PST | wF_PST_M | wPF_PST_M | wNRM | wF2I_M | wP_F2I_M | wAND | wS_AND | wORR | wS_ORR | wXOR | wS_XOR | wLIN | wLIN_M | wP_LIN_M | wLES | wS_LES | wF_LES | wSF_LES | wSHR | wS_SHR | wSRS | wS_SRS | wF_ROT | wF_SU1 | wSF_SU1; end else begin : dec_b2 assign b2 = 1'b0; end endgenerate
 
 // logic for b1
-generate if (ADD | S_ADD | F_ADD | SF_ADD | DIV | S_DIV | F_DIV | SF_DIV | F_SGN | SF_SGN | NEG | F_NEG_M | PF_NEG_M | ABS | F_ABS_M | PF_ABS_M | PST | F_PST_M | PF_PST_M | NRM | I2F_M | P_I2F_M | F2I | ORR | S_ORR | XOR | S_XOR | LAN | S_LAN | LOR | S_LOR | LES | S_LES | F_LES | SF_LES | EQU | S_EQU | SHL | S_SHL | F_ROT | F_SU1 | SF_SU1) begin : dec_b1 assign b1 = wADD | wS_ADD | wF_ADD | wSF_ADD | wDIV | wS_DIV | wF_DIV | wSF_DIV | wF_SGN | wSF_SGN | wNEG | wF_NEG_M | wPF_NEG_M | wABS | wF_ABS_M | wPF_ABS_M | wPST | wF_PST_M | wPF_PST_M | wNRM | wI2F_M | wP_I2F_M | wF2I | wORR | wS_ORR | wXOR | wS_XOR | wLAN | wS_LAN | wLOR | wS_LOR | wLES | wS_LES | wF_LES | wSF_LES | wEQU | wS_EQU | wSHL | wS_SHL | wF_ROT | wF_SU1 | wSF_SU1; end else begin : dec_b1 assign b1 = 1'b0; end endgenerate
+generate if (ADD | S_ADD | F_ADD | SF_ADD | DIV | S_DIV | F_DIV | SF_DIV | F_SGN | SF_SGN | NEG | F_NEG_M | PF_NEG_M | ABS | F_ABS_M | PF_ABS_M | PST | F_PST_M | PF_PST_M | NRM | I2F_M | P_I2F_M | F2I | ORR | S_ORR | XOR | S_XOR | LAN | S_LAN | LOR | S_LOR | LES | S_LES | F_LES | SF_LES | EQU | S_EQU | SHL | S_SHL | F_ROT | F_SU1 | SF_SU1 | XPO | XPO_M) begin : dec_b1 assign b1 = wADD | wS_ADD | wF_ADD | wSF_ADD | wDIV | wS_DIV | wF_DIV | wSF_DIV | wF_SGN | wSF_SGN | wNEG | wF_NEG_M | wPF_NEG_M | wABS | wF_ABS_M | wPF_ABS_M | wPST | wF_PST_M | wPF_PST_M | wNRM | wI2F_M | wP_I2F_M | wF2I | wORR | wS_ORR | wXOR | wS_XOR | wLAN | wS_LAN | wLOR | wS_LOR | wLES | wS_LES | wF_LES | wSF_LES | wEQU | wS_EQU | wSHL | wS_SHL | wF_ROT | wF_SU1 | wSF_SU1 | wXPO | wXPO_M; end else begin : dec_b1 assign b1 = 1'b0; end endgenerate
 
 // logic for b0
-generate if (LOD | P_LOD | LDI | ILI | SET_P | POP | F_INN | PF_INN | F_ADD | SF_ADD | F_MLT | SF_MLT | F_DIV | SF_DIV | SGN | S_SGN | NEG | F_NEG | ABS | F_ABS | PST | F_PST | NRM | I2F | F2I | AND | S_AND | XOR | S_XOR | INV_M | P_INV_M | LOR | S_LOR | LIN_M | P_LIN_M | F_LES | SF_LES | F_GRE | SF_GRE | SHL | S_SHL | SRS | S_SRS | F_SU1 | SF_SU1 | LDA) begin : dec_b0 assign b0 = wLOD | wP_LOD | wLDI | wILI | wSET_P | wPOP | wF_INN | wPF_INN | wF_ADD | wSF_ADD | wF_MLT | wSF_MLT | wF_DIV | wSF_DIV | wSGN | wS_SGN | wNEG | wF_NEG | wABS | wF_ABS | wPST | wF_PST | wNRM | wI2F | wF2I | wAND | wS_AND | wXOR | wS_XOR | wINV_M | wP_INV_M | wLOR | wS_LOR | wLIN_M | wP_LIN_M | wF_LES | wSF_LES | wF_GRE | wSF_GRE | wSHL | wS_SHL | wSRS | wS_SRS | wF_SU1 | wSF_SU1 | wLDA; end else begin : dec_b0 assign b0 = 1'b0; end endgenerate
+generate if (LOD | P_LOD | LDI | ILI | SET_P | POP | F_INN | PF_INN | F_ADD | SF_ADD | F_MLT | SF_MLT | F_DIV | SF_DIV | SGN | S_SGN | NEG | F_NEG | ABS | F_ABS | PST | F_PST | NRM | I2F | F2I | AND | S_AND | XOR | S_XOR | INV_M | P_INV_M | LOR | S_LOR | LIN_M | P_LIN_M | F_LES | SF_LES | F_GRE | SF_GRE | SHL | S_SHL | SRS | S_SRS | F_SU1 | SF_SU1 | F_SCL | SF_SCL | XPO_M | LDA) begin : dec_b0 assign b0 = wLOD | wP_LOD | wLDI | wILI | wSET_P | wPOP | wF_INN | wPF_INN | wF_ADD | wSF_ADD | wF_MLT | wSF_MLT | wF_DIV | wSF_DIV | wSGN | wS_SGN | wNEG | wF_NEG | wABS | wF_ABS | wPST | wF_PST | wNRM | wI2F | wF2I | wAND | wS_AND | wXOR | wS_XOR | wINV_M | wP_INV_M | wLOR | wS_LOR | wLIN_M | wP_LIN_M | wF_LES | wSF_LES | wF_GRE | wSF_GRE | wSHL | wS_SHL | wSRS | wS_SRS | wF_SU1 | wSF_SU1 | wF_SCL | wSF_SCL | wXPO_M | wLDA; end else begin : dec_b0 assign b0 = 1'b0; end endgenerate
 
 // combine logic into ula_op
 always @ (posedge clk) ula_op <= {b5,b4,b3,b2,b1,b0};
@@ -495,4 +506,8 @@ endmodule
     99: ula_op  <= 6'd48;    //  F_SU2   -> floating-point subtraction with memory at input 2
    100: ula_op  <= 6'd47;    // SF_SU1   -> floating-point subtraction with stack at input 1
    101: ula_op  <= 6'd48;    // SF_SU2   -> floating-point subtraction with stack at input 2
+   104: ula_op  <= 6'd49;    //  F_SCL   -> scale float by 2^k, k from memory
+   105: ula_op  <= 6'd49;    // SF_SCL   -> scale float by 2^k, k from stack
+   106: ula_op  <= 6'd50;    //    XPO   -> base-2 exponent of float (acc) as int
+   107: ula_op  <= 6'd51;    //  XPO_M   -> base-2 exponent of float (memory) as int
     ------------------------------------------------*/